+++ integration-tests/src/test/java/com/transferwise/tasks/triggering/KafkaTasksExecutionTriggererIntTest.java
@@ -116,11 +116,11 @@ class KafkaTasksExecutionTriggererIntTest extends BaseIntTest {
   }
 
   static class TestPartitionKeyStrategy implements IPartitionKeyStrategy {
-    private static final UUID key = UUID.fromString(PARTITION_KEY);
-    
+    private static final UUID KEY = UUID.fromString(PARTITION_KEY);
+
     @Override
     public String createPartitionKey(BaseTask task) {
-      return key.toString();
+      return KEY.toString();
     }
   }
 }